x64: xf16/fp8 kernels: eliminate implicit narrowing conversions

+++ src/cpu/x64/jit_avx512_core_fp8cvt.cpp
@@ -46,16 +46,16 @@ void fp8_conversion_e5m2_t::prepare_table() {
             case 2: index = i + 1; break;
             case 3: index = i - 2; break;
         }
-        host_->db(index);
+        host_->db(static_cast<int>(index));
     }
     // Data for bf16
     for (size_t i = 0; i < 32; ++i) {
         size_t index = 4 * (i / 2) + (i % 2);
-        host_->db(index);
+        host_->db(static_cast<int>(index));
     }
     for (size_t i = 32; i < 64; ++i) {
         size_t index = 4 * ((i - 32) / 2) + (i % 2) + 2;
-        host_->db(index);
+        host_->db(static_cast<int>(index));
     }
 
     // Other tables are not needed if fp8 is native
@@ -87,11 +87,11 @@ void fp8_conversion_e4m3_t::prepare_table() {
     host_->L(label_vnni_permute_index_table_);
     for (size_t i = 0; i < 32; ++i) {
         size_t index = 4 * (i / 2) + (i % 2);
-        host_->db(index);
+        host_->db(static_cast<int>(index));
     }
     for (size_t i = 32; i < 64; ++i) {
         size_t index = 4 * ((i - 32) / 2) + (i % 2) + 2;
-        host_->db(index);
+        host_->db(static_cast<int>(index));
     }
 
     host_->L(label_table_from_f8_);
